Restructure plugin files and update readme
