+++ src/game_sdk/hosted_game/sdk.py
@@ -76,6 +76,7 @@ class GameSDK:
             payload["tweetId"] = tweet_id
             
         print(payload)
+
         response = requests.post(
             url,
             json={
